added tests to test the accuracy of the BiCGStab solver.

+++ CuDenseBiCGStab/Solvers/Solver.cuh
@@ -22,6 +22,6 @@ public:
 	// Method to get the number of iterations performed
 	virtual int getNumIterations() const = 0;
 	// Method to get the residual norm
-	virtual float getResidualNorm() const = 0;
+	virtual double getResidualNorm() const = 0;
 	virtual void setOperator(const LinearOperator<T>& A) = 0;
 };
